+++ .storybook/main.js
@@ -2,8 +2,6 @@ module.exports = {
   stories: ["../src/**/*.stories.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
   addons: [
     "@storybook/addon-links",
-    "@storybook/addon-essentials",
-    "@storybook/preset-scss",
     {
       name: "@storybook/addon-docs",
       options: {
@@ -13,6 +11,8 @@ module.exports = {
         transcludeMarkdown: true,
       },
     },
+    "@storybook/addon-essentials",
+    "@storybook/preset-scss",
   ],
   typescript: {
     reactDocgen: "react-docgen-typescript",